+++ src/app/api/applications/route.ts
@@ -2,12 +2,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-import axios from 'axios';
 import { authOptions, ExtendedSession } from '@/utils/auth';
 import { getServerSession } from 'next-auth';
 import { NextResponse } from 'next/server';
-import serverConfig from '@/config/serverConfig';
-import { decrypt } from '@/utils/encryption';
+import { createApplication } from '@/services/daam/index.server';
 
 export async function POST(request: Request) {
   const session: ExtendedSession | null = await getServerSession(authOptions);
@@ -21,16 +19,7 @@ export async function POST(request: Request) {
       return NextResponse.json({ error: 'datasetIds are required' }, { status: 400 });
     }
 
-    const requestBody = {
-      datasetIds: datasetIds,
-    };
-
-    await axios.post(`${serverConfig.daamUrl}/api/v1/applications/create-application`, requestBody, {
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${decrypt(session.access_token)}`,
-      },
-    });
+    await createApplication(datasetIds, session);
 
     return NextResponse.json({ success: true });
   } catch (error) {
